Add destroy scripts
