+++ installer/cli/uninstallcli.py
@@ -1,5 +1,6 @@
 import click
 from installer.helpers.destroyprep import destroyprep
+from installer.helpers.terraform import exec_terraform_destroy, get_terraform_output_var
 
 
 @click.group()
@@ -15,8 +16,18 @@ def uninstall(mode):
     click.secho('\n\nConfiguring uninstall', fg='blue')
     click.secho('\nLogging to `uninstall.log`', fg='green')
 
+    click.secho('\n\n Collecting terraform output vars from existing install...', fg='blue')
+    stackprefix = get_terraform_output_var("jazzprefix")
+    identity = get_terraform_output_var("jazzusername")
+
+    click.secho(
+        '\nRunning destroy prep with option: {0}, prefix: {1}, and identity: {2}'
+        .format(mode, stackprefix, identity), fg='green')
+
     if mode == "all":
-        destroyprep(True)
+        destroyprep(stackprefix, identity, True)
     else:
-        destroyprep(False)
-    # update_main_terraform_vars(branch, adminemail, stackprefix, region, tags)
+        destroyprep(stackprefix, identity, False)
+
+    click.secho('\nRunning terraform destroy', fg='green')
+    exec_terraform_destroy()
