+++ installer/helpers/destroy/delete_platform_services.py
@@ -1,6 +1,4 @@
-#!/usr/bin/python3
 import json
-import sys
 import subprocess
 import os
 import os.path
@@ -45,41 +43,24 @@ def deleteCloudFormationService(service_name):
         print('Error service not found::' + service_name)
 
 
-if (len(sys.argv) != 3):
-    print(
-        "Error - Please provide Argument1='stack name' and Argument1='Delete Client servies option true or false' "
-    )
-    print("Argument1=<StackName> Argument2=<true/false>")
-    exit(1)
+def delete_platform_services(stackname, deleteClientServices=False):
+    print("\r\nStarting deletion of micro services\r\n\r\n")
 
-print(str(sys.argv))
+    # Delete user services Cloud formations
+    fname = 'listservice.json'
+    if os.path.isfile(fname):
+        os.remove(fname)
 
-stackName = sys.argv[1].lower() + "-"
-deleteClientServices = sys.argv[2]
+    valresp = getServicesList()
+    if (valresp != 0):
+        exit(1)
 
-if (deleteClientServices.lower() != 'true'):
-    exit(0)
+    jsonFile = open(fname, 'r')
+    values = json.load(jsonFile)
+    jsonFile.close()
 
-print("\r\nStarting deletion of micro services\r\n\r\n")
+    ss = values['StackSummaries']
 
-# Delete user services Cloud formations
-fname = 'listservice.json'
-if os.path.isfile(fname):
-    os.remove(fname)
-
-valresp = getServicesList()
-if (valresp != 0):
-    exit(1)
-
-jsonFile = open(fname, 'r')
-values = json.load(jsonFile)
-jsonFile.close()
-
-ss = values['StackSummaries']
-
-spLen = len(stackName)
-
-inval = 0
-for item in ss:
-    if (item['StackName'].startswith(stackName)):
-        deleteCloudFormationService(item['StackName'])
+    for item in ss:
+        if (item['StackName'].startswith(stackname)):
+            deleteCloudFormationService(item['StackName'])
